feat(ext): add Indigo enum with shades and related methods
- refactor(ext): remove redundant missing_docs attributes in Blue enum
- refactor(ext): move Indigo enum to its own module and update mod.rs to include it
